(+) Memperbaiki Migration

+++ database/migrations/2025_02_25_080725_create_m_barang_table.php
@@ -10,18 +10,20 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('m_barang', function (Blueprint $table) {
-            $table->id('barang_id');
-            $table->foreignId('kategori_id')->constrained('m_kategori');
-            $table->string('barang_kode', 10);
-            $table->string('barang_nama', 100);
-            $table->integer('harga_beli');
-            $table->integer('harga_jual');
-            $table->timestamps();
+        if (!Schema::hasTable('m_barang')) {
+            Schema::create('m_barang', function (Blueprint $table) {
+                $table->id('barang_id'); // bigint(20) unsigned
+                $table->unsignedBigInteger('kategori_id'); // bigint(20) unsigned
+                $table->string('barang_kode', 10); // varchar(10)
+                $table->string('barang_nama', 100); // varchar(100)
+                $table->integer('harga_beli'); // int(11)
+                $table->integer('harga_jual'); // int(11)
+                $table->timestamps(); // created_at, updated_at
 
-            $table->foreign('kategori_id')->references('kategori_id')->on('m_kategori')->onDelete('cascade');
-
-        });
+                // Menambahkan foreign key
+                $table->foreign('kategori_id')->references('kategori_id')->on('m_kategori')->onDelete('cascade');
+            });
+        }
     }
 
     /**
